Refactor to move HDF object code to their own packages
