+++ org.eclipse.dawnsci.hdf5/META-INF/MANIFEST.MF
@@ -14,8 +14,10 @@ Require-Bundle: ncsa.hdf;bundle-version="[2.9.0,3.0.0)",
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Bundle-ActivationPolicy: lazy
 Bundle-ClassPath: .
-Export-Package: org.eclipse.dawnsci.hdf5,
- org.eclipse.dawnsci.hdf5.nexus,
- org.eclipse.dawnsci.hdf5.operation
+Export-Package: org.eclipse.dawnsci.hdf.object,
+ org.eclipse.dawnsci.hdf.object.nexus,
+ org.eclipse.dawnsci.hdf.object.operation,
+ org.eclipse.dawnsci.hdf5,
+ org.eclipse.dawnsci.hdf5.nexus
 Import-Package: org.slf4j;version="[1.7.2,2.0.0)"
 Service-Component: OSGI-INF/*.xml
