+++ org.eclipse.dawnsci.hdf5/src/org/eclipse/dawnsci/hdf/object/H5Utils.java
@@ -9,7 +9,7 @@
  * Contributors:
  *    Matthew Gerring - initial API and implementation and/or initial documentation
  *******************************************************************************/ 
-package org.eclipse.dawnsci.hdf5;
+package org.eclipse.dawnsci.hdf.object;
 
 import ncsa.hdf.object.Datatype;
 import ncsa.hdf.object.h5.H5Datatype;
