+++ org.eclipse.dawnsci.hdf5/src/org/eclipse/dawnsci/hdf/object/HierarchicalDataFactory.java
@@ -9,7 +9,7 @@
  * Contributors:
  *    Matthew Gerring - initial API and implementation and/or initial documentation
  *******************************************************************************/ 
-package org.eclipse.dawnsci.hdf5;
+package org.eclipse.dawnsci.hdf.object;
 
 import java.io.File;
 import java.io.IOException;
@@ -21,6 +21,7 @@ import java.util.concurrent.locks.ReentrantLock;
 
 import ncsa.hdf.object.FileFormat;
 
+import org.eclipse.dawnsci.hdf5.HDF5FileFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
