+++ org.eclipse.dawnsci.hdf5/src/org/eclipse/dawnsci/hdf/object/nexus/IFindInNexus.java
@@ -9,19 +9,18 @@
  * Contributors:
  *    Matthew Gerring - initial API and implementation and/or initial documentation
  *******************************************************************************/
-package org.eclipse.dawnsci.hdf5;
+package org.eclipse.dawnsci.hdf.object.nexus;
 
-import ncsa.hdf.object.FileFormat;
 
 /**
- * Internal use only. Do not use this interface externally to the plugin
- * TODO Create an internal package and move interface to it.
- * 
- * @author Matthew Gerring
- *
+ * This interface describes a single method that returns a boolean when given a
+ * Hierarchical object. This allows classes to be created specifically to test
+ * the type/name/contents of a Hierarchical object, and return whether the object
+ * matches or not. To be used with methods that iterate over Hierarchical files to 
+ * locate specific objects.
  */
-public interface IFileFormatDataFile extends IHierarchicalDataFile {
-
-	public FileFormat getFileFormat();
-
+public interface IFindInNexus
+{
+    boolean inNexus(String nexusObjectPath);
+    
 }
