+++ org.eclipse.dawnsci.hdf5/src/org/eclipse/dawnsci/hdf/object/nexus/NexusFindDatasetByName.java
@@ -9,10 +9,10 @@
  * Contributors:
  *    Matthew Gerring - initial API and implementation and/or initial documentation
  *******************************************************************************/
-package org.eclipse.dawnsci.hdf5.nexus;
+package org.eclipse.dawnsci.hdf.object.nexus;
 
 
-import org.eclipse.dawnsci.hdf5.IHierarchicalDataFile;
+import org.eclipse.dawnsci.hdf.object.IHierarchicalDataFile;
 
 public class NexusFindDatasetByName implements IFindInNexus {
